+++ app/Helpers/Delete/Models/DeleteHousingFile.php
@@ -10,12 +10,10 @@ namespace App\Helpers\Delete\Models;
 
 use App\Helpers\Delete\DeleteInterface;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 /**
  * Class DeleteHousingFile
- *
- * Relation: 1-n Documents. Action: dissociate
- * Relation: 1-n Tasks & notes. Action: call DeleteTask
  *
  * @package App\Helpers\Delete
  */
@@ -34,19 +32,42 @@ class DeleteHousingFile implements DeleteInterface
         $this->housingFile = $housingFile;
     }
 
-    /** Main method for deleting this model and all it's relations
+    /** If it's called by the cleanup functionality, we land on this function, else on the delete function
      *
      * @return array
      * @throws
      */
+    public function cleanup()
+    {
+        try{
+            return $this->delete();
+        }catch (\Exception $exception){
+            Log::error('Fout bij opschonen Woningdossiers', [
+                'exception' => $exception->getMessage(),
+                'errormessages' => implode(' | ', $this->errorMessage),
+            ]);
+            array_push($this->errorMessage, "Fout bij opschonen Woningdossiers. (meld dit bij Econobis support)");
+            return $this->errorMessage;
+        }
+    }
+
+    /** Main method for deleting this model and all it's relations
+     *
+     * @return array errorMessage array
+     * @throws
+     */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->housingFile->delete();
+        if( count($this->errorMessage) === 0 ) {
+            $this->housingFile->delete();
+        }
 
         return $this->errorMessage;
     }
@@ -56,7 +77,8 @@ class DeleteHousingFile implements DeleteInterface
      */
     public function canDelete()
     {
-
+        // van hier uit altijd true
+        return true;
     }
 
     /** Deletes models recursive
@@ -66,15 +88,15 @@ class DeleteHousingFile implements DeleteInterface
     {
         foreach ($this->housingFile->housingFileSpecifications as $specification) {
             $deleteSpecifiction = new DeleteHousingFileSpecification($specification);
-            $this->errorMessage = array_merge($this->errorMessage, $deleteSpecifiction->delete());
+            $this->errorMessage = array_merge($this->errorMessage, ( $deleteSpecifiction->delete() ?? [] ) );
         }
         foreach ($this->housingFile->tasks as $task) {
             $deleteTask = new DeleteTask($task);
-            $this->errorMessage = array_merge($this->errorMessage, $deleteTask->delete());
+            $this->errorMessage = array_merge($this->errorMessage, ( $deleteTask->delete() ?? [] ) );
         }
         foreach ($this->housingFile->notes as $task) {
             $deleteTask = new DeleteTask($task);
-            $this->errorMessage = array_merge($this->errorMessage, $deleteTask->delete());
+            $this->errorMessage = array_merge($this->errorMessage, ( $deleteTask->delete() ?? [] ) );
         }
     }
