+++ app/Helpers/Delete/Models/DeleteIntake.php
@@ -8,17 +8,12 @@
 
 namespace App\Helpers\Delete\Models;
 
-
 use App\Helpers\Delete\DeleteInterface;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 /**
  * Class DeleteIntake
- *
- * Relation: 1-n Documents. Action: dissociate
- * Relation: 1-n Emails. Action: dissociate
- * Relation: 1-n Opportunities. Action: call DeleteOpportunity
- * Relation: 1-n Tasks & notes. Action: call DeleteTask
  *
  * @package App\Helpers\Delete\Models
  */
@@ -36,6 +31,25 @@ class DeleteIntake implements DeleteInterface
         $this->intake = $intake;
     }
 
+    /** If it's called by the cleanup functionality, we land on this function, else on the delete function
+     *
+     * @return array
+     * @throws
+     */
+    public function cleanup()
+    {
+        try{
+            return $this->delete();
+        }catch (\Exception $exception){
+            Log::error('Fout bij opschonen Intakes', [
+                'exception' => $exception->getMessage(),
+                'errormessages' => implode(' | ', $this->errorMessage),
+            ]);
+            array_push($this->errorMessage, "Fout bij opschonen Intakes. (meld dit bij Econobis support)");
+            return $this->errorMessage;
+        }
+    }
+
     /** Main method for deleting this model and all it's relations
      *
      * @return array errorMessage array
@@ -43,12 +57,16 @@ class DeleteIntake implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->intake->delete();
+        if( count($this->errorMessage) === 0 ) {
+            $this->intake->delete();
+        }
 
         return $this->errorMessage;
     }
@@ -56,17 +74,21 @@ class DeleteIntake implements DeleteInterface
     /** Checks if the model can be deleted and sets error messages
      *
      */
-    public function canDelete()
+    public function canDelete(): bool
     {
+        $intakeAddress = $this->intake->address?->fullAddress ?? '*adres onbekend*';
         // 25-04-2024: Verwijderen 1 voor 1 mag ook ongeacht de status van de intake
 //        if(!($this->intake->intake_status_id === 2 || $this->intake->intake_status_id === 3)){
 //            array_push($this->errorMessage, "Er is nog een intake die niet gesloten is.");
+//            return false;
 //        }
         // 25-04-2024: Verwijderen mag niet meer als er nog kansen onder hangen
         if($this->intake->opportunities()->count() > 0){
-            array_push($this->errorMessage, "Onder intake " . ($this->intake->address ? $this->intake->address->fullAddress : '*adres onbekend*') . " hangen nog kansen, verwijderen intake niet mogelijk.");
+            array_push($this->errorMessage, "Onder intake (" . $this->intake->id . ") " . $intakeAddress . " hangen nog kansen, verwijderen intake niet mogelijk.");
+            return false;
         }
 
+        return true;
     }
 
     /** Deletes models recursive
@@ -75,18 +97,18 @@ class DeleteIntake implements DeleteInterface
     {
         foreach ($this->intake->tasks as $task){
             $deleteTask = new DeleteTask($task);
-            $this->errorMessage = array_merge($this->errorMessage, $deleteTask->delete());
+            $this->errorMessage = array_merge($this->errorMessage, ( $deleteTask->delete() ?? [] ) );
         }
 
         foreach ($this->intake->notes as $note){
             $deleteTask = new DeleteTask($note);
-            $this->errorMessage = array_merge($this->errorMessage, $deleteTask->delete());
+            $this->errorMessage = array_merge($this->errorMessage, ( $deleteTask->delete() ?? [] ) );
         }
 
         // 25-04-2024: Verwijderen mag niet meer als er nog kansen onder hangen
 //        foreach ($this->intake->opportunities as $opportunity){
 //            $deleteOpportunity = new DeleteOpportunity($opportunity);
-//            $this->errorMessage = array_merge($this->errorMessage, $deleteOpportunity->delete());
+//            $this->errorMessage = array_merge($this->errorMessage, ( $deleteOpportunity->delete() ?? [] ) );
 //        }
     }
 
@@ -103,6 +125,9 @@ class DeleteIntake implements DeleteInterface
             $document->intake()->dissociate();
             $document->save();
         }
+
+        $this->intake->reasons()->detach();
+        $this->intake->sources()->detach();
     }
 
     /**
