+++ app/Helpers/Delete/Models/DeleteQuotationRequest.php
@@ -8,15 +8,12 @@
 
 namespace App\Helpers\Delete\Models;
 
-
 use App\Helpers\Delete\DeleteInterface;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 /**
  * Class DeleteQuotationRequest
- *
- * Relation: 1-n Emails. Action: dissociate
- * Relation: 1-n Documents. Action: dissociate
  *
  * @package App\Helpers\Delete\Models
  */
@@ -29,10 +26,28 @@ class DeleteQuotationRequest implements DeleteInterface
      *
      * @param Model $quotationRequest the model to delete
      */
-
     public function __construct(Model $quotationRequest)
     {
         $this->quotationRequest = $quotationRequest;
+    }
+
+    /** If it's called by the cleanup functionality, we land on this function, else on the delete function
+     *
+     * @return array
+     * @throws
+     */
+    public function cleanup()
+    {
+        try{
+            return $this->delete();
+        }catch (\Exception $exception){
+            Log::error('Fout bij opschonen Kansacties', [
+                'exception' => $exception->getMessage(),
+                'errormessages' => implode(' | ', $this->errorMessage),
+            ]);
+            array_push($this->errorMessage, "Fout bij opschonen Kansacties. (meld dit bij Econobis support)");
+            return $this->errorMessage;
+        }
     }
 
     /** Main method for deleting this model and all it's relations
@@ -42,12 +57,16 @@ class DeleteQuotationRequest implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->quotationRequest->delete();
+        if( count($this->errorMessage) === 0 ) {
+            $this->quotationRequest->delete();
+        }
 
         return $this->errorMessage;
     }
@@ -56,19 +75,8 @@ class DeleteQuotationRequest implements DeleteInterface
      */
     public function canDelete()
     {
-        // Rewrite if to in_array check
-//        if(!in_array($this->quotationRequest->status_id, [2, 3, 4, 7, 8])) {
-//            if ($this->quotationRequest->status_id === 9) {
-//                array_push($this->errorMessage, "Een bezoek met status \"Afspraak uitgevoerd\" kan niet verwijderd worden.");
-//            } else {
-//                array_push($this->errorMessage, "Er is nog een open offerteverzoek.");
-//            }
-//        }
-        //
-        // 25-04-2024: Verwijderen 1 voor 1 mag ook ongeacht de status van de kansactie
-//        if($this->quotationRequest->status->is_pending_status){
-//            array_push($this->errorMessage, "Er is nog een " . ($this->quotationRequest->opportunityAction ? $this->quotationRequest->opportunityAction->name : "onbekend") . " met een onderhanden status " . ($this->quotationRequest->status ? $this->quotationRequest->status->name : "onbekend" ) );
-//        }
+        // van hieruit altijd true
+        return true;
     }
 
     /** Deletes models recursive
